+++ sdk/storage/azure-storage-blobs/src/blob_responses.cpp
@@ -4,6 +4,9 @@
 #include "azure/storage/blobs/blob_responses.hpp"
 
 #include "azure/storage/blobs/blob_client.hpp"
+#include "azure/storage/blobs/blob_container_client.hpp"
+#include "azure/storage/blobs/blob_service_client.hpp"
+#include "azure/storage/blobs/page_blob_client.hpp"
 
 namespace Azure { namespace Storage { namespace Blobs {
 
@@ -56,6 +59,52 @@ namespace Azure { namespace Storage { namespace Blobs {
 
       std::this_thread::sleep_for(period);
     };
+  }
+
+  void ListBlobContainersPagedResponse::OnNextPage(const Azure::Core::Context& context)
+  {
+    m_operationOptions.ContinuationToken = NextPageToken;
+    *this = m_blobServiceClient->ListBlobContainers(m_operationOptions, context);
+  }
+
+  void FindBlobsByTagsPagedResponse::OnNextPage(const Azure::Core::Context& context)
+  {
+    m_operationOptions.ContinuationToken = NextPageToken;
+    *this = m_blobServiceClient->FindBlobsByTags(
+        m_tagFilterSqlExpression, m_operationOptions, context);
+  }
+
+  void ListBlobsPagedResponse::OnNextPage(const Azure::Core::Context& context)
+  {
+    m_operationOptions.ContinuationToken = NextPageToken;
+    *this = m_blobContainerClient->ListBlobs(m_operationOptions, context);
+  }
+
+  void ListBlobsByHierarchyPagedResponse::OnNextPage(const Azure::Core::Context& context)
+  {
+    m_operationOptions.ContinuationToken = NextPageToken;
+    *this = m_blobContainerClient->ListBlobsByHierarchy(m_delimiter, m_operationOptions, context);
+  }
+
+  void GetPageRangesPagedResponse::OnNextPage(const Azure::Core::Context& context)
+  {
+    *this = m_pageBlobClient->GetPageRanges(m_operationOptions, context);
+  }
+
+  void GetPageRangesDiffPagedResponse::OnNextPage(const Azure::Core::Context& context)
+  {
+    if (m_previousSnapshot.HasValue())
+    {
+      *this = m_pageBlobClient->GetPageRangesDiff(
+          m_previousSnapshot.Value(), m_operationOptions, context);
+    }
+    else if (m_previousSnapshotUrl.HasValue())
+    {
+      *this = m_pageBlobClient->GetManagedDiskPageRangesDiff(
+          m_previousSnapshotUrl.Value(), m_operationOptions, context);
+    }
+    // Execution is never expected to reach here
+    std::abort();
   }
 
 }}} // namespace Azure::Storage::Blobs