+++ sdk/storage/azure-storage-files-shares/src/share_client.cpp
@@ -272,31 +272,4 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
     return Azure::Response<std::string>(result.Value.FilePermission, std::move(result.RawResponse));
   }
 
-  Azure::Response<Models::ListFilesAndDirectoriesSinglePageResult>
-  ShareClient::ListFilesAndDirectoriesSinglePage(
-      const ListFilesAndDirectoriesSinglePageOptions& options,
-      const Azure::Core::Context& context) const
-  {
-    auto protocolLayerOptions
-        = _detail::ShareRestClient::Directory::ListFilesAndDirectoriesSinglePageOptions();
-    protocolLayerOptions.Prefix = options.Prefix;
-    protocolLayerOptions.ContinuationToken = options.ContinuationToken;
-    protocolLayerOptions.MaxResults = options.PageSizeHint;
-    auto result = _detail::ShareRestClient::Directory::ListFilesAndDirectoriesSinglePage(
-        m_shareUrl, *m_pipeline, context, protocolLayerOptions);
-    Models::ListFilesAndDirectoriesSinglePageResult ret;
-    ret.ServiceEndpoint = std::move(result.Value.ServiceEndpoint);
-    ret.ShareName = std::move(result.Value.ShareName);
-    ret.ShareSnapshot = std::move(result.Value.ShareSnapshot);
-    ret.DirectoryPath = std::move(result.Value.DirectoryPath);
-    ret.Prefix = std::move(result.Value.Prefix);
-    ret.PageSizeHint = result.Value.PageSizeHint;
-    ret.ContinuationToken = std::move(result.Value.ContinuationToken);
-    ret.DirectoryItems = std::move(result.Value.SinglePage.DirectoryItems);
-    ret.FileItems = std::move(result.Value.SinglePage.FileItems);
-
-    return Azure::Response<Models::ListFilesAndDirectoriesSinglePageResult>(
-        std::move(ret), std::move(result.RawResponse));
-  }
-
 }}}} // namespace Azure::Storage::Files::Shares